spi: Add support for the Aspeed ast2500 SPI controllers
The Aspeed AST2500 SoC comes with three static memory controllers, all
with a similar interface :

 * Firmware SPI Memory Controller (FMC)
   . BMC firmware
   . 3 chip select pins (CE0 ~ CE2)
   . supports SPI type flash memory (CE0 ~ CE1)
   . CE2 can be of NOR type flash but this is not supported by the
     driver

 * SPI Flash Controller (SPI1 and SPI2)
   . host firmware
   . 2 chip select pins (CE0 ~ CE1)

Each controller has a defined AHB window for its registers and another
AHB window on which all the flash devices are mapped. Each device is
assigned a memory range through a set of registers called the Segment
Address Registers.

Devices are controlled using two different modes: the USER command
mode or the READ/WRITE command mode. When in USER command mode,
accesses to the AHB window of the SPI flash device are translated into
SPI command transfers. When in READ/WRITE command mode, the HW
generates the SPI commands depending on the setting of the CE control
register.

The following driver supports the FMC and the SPI controllers with the
attached SPI flash devices. When the controller is probed, the driver
performs a read timing calibration using specific DMA control
registers (FMC only). The driver's primary goal is to support the
first device of the FMC controller on which reside U-Boot but it
should support the other controllers also.

The Aspeed FMC controller automatically detects at boot time if a
flash device is in 4BYTE address mode and self configures to use the
appropriate address width. This can be a problem for the U-Boot SPI
Flash layer which only supports 3 byte addresses. In such a case, a
warning is emitted and the address width is fixed when sent on the
bus.

Signed-off-by: Cédric Le Goater <clg@kaod.org>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Cédric Le Goater <clg@kaod.org>